refactor(verifier): extract writeTrajectoryDir + shouldPersistTrajectory
Lift the on-disk persistence helpers from TrajectoryRecorder into
verifier/trajectory.ts so #2137's harness adapter can share them. Also
drop the recorder's no-op .replace("T","T") and the WHAT-narration
comments per project policy.
